pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "ctf-agent"
version = "0.1.0"
description = "AI-Powered CTF Auto-Solver Agent Framework"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "xunzhang85"},
]
keywords = ["ctf", "agent", "ai", "cybersecurity", "llm", "security"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]

dependencies = [
    "openai>=1.0.0",
    "anthropic>=0.20.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "click>=8.0",
    "httpx>=0.25.0",
    "pydantic>=2.0",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.5",
    "pre-commit>=3.0",
]
tools = [
    "pwntools>=4.10",
    "requests>=2.28",
    "beautifulsoup4>=4.12",
    "lxml>=4.9",
    "paramiko>=3.0",
]

[project.scripts]
agent = "agent.cli:main"

[project.urls]
Homepage = "https://github.com/xunzhang85/Agent"
Documentation = "https://xunzhang85.github.io/Agent/"
Repository = "https://github.com/xunzhang85/Agent"
Issues = "https://github.com/xunzhang85/Agent/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
